Move Microtimer support to hal.cpp
